+++ tests/test_bookmarks.py
@@ -2,6 +2,7 @@ import os
 from time import perf_counter
 
 import singer
+import base
 import tap_tester.connections as connections
 import tap_tester.menagerie   as menagerie
 import tap_tester.runner      as runner
@@ -13,9 +14,6 @@ LOGGER = singer.get_logger()
 
 class TestSquareIncrementalReplication(TestSquareBaseParent.TestSquareBase):
 
-    @staticmethod
-    def name():
-        return "tap_tester_square_incremental_replication"
 
     def testable_streams_dynamic(self):
         return self.dynamic_data_streams().difference(self.untestable_streams())
@@ -34,7 +32,7 @@ class TestSquareIncrementalReplication(TestSquareBaseParent.TestSquareBase):
 
     @classmethod
     def tearDownClass(cls):
-        print("\n\nTEST TEARDOWN\n\n")
+        LOGGER.info('\n\nTEST TEARDOWN\n\n')
 
     def run_sync(self, conn_id):
         """
@@ -61,15 +59,17 @@ class TestSquareIncrementalReplication(TestSquareBaseParent.TestSquareBase):
 
         self.START_DATE = self.get_properties().get('start_date')
 
-        print("\n\nTESTING WITH DYNAMIC DATA IN SQUARE_ENVIRONMENT: {}".format(os.getenv('TAP_SQUARE_ENVIRONMENT')))
+        LOGGER.info('\n\nTESTING WITH DYNAMIC DATA IN SQUARE_ENVIRONMENT: {}'.format(os.getenv('TAP_SQUARE_ENVIRONMENT')))
         self.bookmarks_test(self.testable_streams_dynamic().intersection(self.sandbox_streams()))
 
+        TestSquareBaseParent.TestSquareBase.test_name = self.TEST_NAME_PROD
         self.set_environment(self.PRODUCTION)
         production_testable_streams = self.testable_streams_dynamic().intersection(self.production_streams())
 
         if production_testable_streams:
-            print("\n\nTESTING WITH DYNAMIC DATA IN SQUARE_ENVIRONMENT: {}".format(os.getenv('TAP_SQUARE_ENVIRONMENT')))
+            LOGGER.info('\n\nTESTING WITH DYNAMIC DATA IN SQUARE_ENVIRONMENT: {}'.format(os.getenv('TAP_SQUARE_ENVIRONMENT')))
             self.bookmarks_test(production_testable_streams)
+        TestSquareBaseParent.TestSquareBase.test_name = self.TEST_NAME_SANDBOX
 
     def bookmarks_test(self, testable_streams):
         """
@@ -84,13 +84,18 @@ class TestSquareIncrementalReplication(TestSquareBaseParent.TestSquareBase):
         For EACH stream that is incrementally replicated there are multiple rows of data with
             different values for the replication key
         """
-        print("\n\nRUNNING {}\n\n".format(self.name()))
+        LOGGER.info('\n\nRUNNING {}_bookmark\n\n'.format(self.name()))
 
+        testable_streams = testable_streams - {'payments'}
+        # Fail the test when the JIRA card is done to allow stream to be re-added and tested
+        self.assertNotEqual(base.JIRA_CLIENT.get_status_category('TDL-26905'), 
+                         'done',
+                         msg='JIRA ticket has moved to done, re-add the payments stream to the testable streams')
         # Ensure tested streams have existing records
         expected_records_first_sync = self.create_test_data(testable_streams, self.START_DATE, force_create_records=True)
 
         # Instantiate connection with default start
-        conn_id = connections.ensure_connection(self)
+        conn_id = connections.ensure_connection(self, payload_hook=self.preserve_access_token)
 
         # run in check mode
         check_job_name = runner.run_check_mode(self, conn_id)
@@ -176,15 +181,6 @@ class TestSquareIncrementalReplication(TestSquareBaseParent.TestSquareBase):
 
                 if not first_rec:
                     raise RuntimeError("Unable to find any any orders with state other than COMPLETED")
-            elif stream == 'roles':  # Use the first available role that has limited permissions (where is_owner = False)
-                for message in first_sync_records.get(stream).get('messages'):
-                    data = message.get('data')
-                    if not data['is_owner'] and 'role' in data['name']:
-                        first_rec = message.get('data')
-                        break
-
-                if not first_rec:
-                    raise RuntimeError("Unable to find any any orders with state other than COMPLETED")
             else: # By default we want the last created record
                 last_message = first_sync_records.get(stream).get('messages')[-1]
                 if last_message.get('data') and not last_message.get('data').get('is_deleted'):
@@ -258,7 +254,6 @@ class TestSquareIncrementalReplication(TestSquareBaseParent.TestSquareBase):
             assert updated_record, "Failed to update a {} record".format('payments')
             assert len(updated_record) == 1, "Updated too many {} records".format('payments')
 
-            expected_records_second_sync['payments'] += updated_record[0]
             updated_records['payments'] += updated_record[0]
 
         # adjust expectations for full table streams to include the expected records from sync 1
@@ -277,9 +272,9 @@ class TestSquareIncrementalReplication(TestSquareBaseParent.TestSquareBase):
         # Adjust expectations for datetime format
         for record_desc, records in [("created", created_records), ("updated", updated_records),
                                      ("2nd sync expected records", expected_records_second_sync)]:
-            print("Adjusting epxectations for {} records".format(record_desc))
+            LOGGER.info('Adjusting epxectations for {} records'.format(record_desc))
             for stream, expected_records in records.items():
-                print("\tadjusting for stream: {}".format(stream))
+                LOGGER.info('\tadjusting for stream: {}'.format(stream))
                 self.modify_expected_records(expected_records)
 
         # ensure validity of expected_records_second_sync
@@ -289,7 +284,7 @@ class TestSquareIncrementalReplication(TestSquareBaseParent.TestSquareBase):
                     self.assertEqual(1, len(expected_records_second_sync.get(stream)),
                                      msg="Expectations are invalid for incremental stream {}".format(stream))
                 elif stream == 'orders': # ORDERS are returned inclusive on the datetime queried
-                    self.assertEqual(3, len(expected_records_second_sync.get(stream)),
+                    self.assertGreaterEqual(3, len(expected_records_second_sync.get(stream)),
                                      msg="Expectations are invalid for incremental stream {}".format(stream))
                 else:  # Most streams will have 2 records from the Update and Insert
                     self.assertEqual(2, len(expected_records_second_sync.get(stream)),
@@ -328,7 +323,7 @@ class TestSquareIncrementalReplication(TestSquareBaseParent.TestSquareBase):
         PARENT_FIELD_MISSING_SUBFIELDS = {'payments': {'card_details'}}
 
         # BUG_2 | https://stitchdata.atlassian.net/browse/SRCE-5143
-        MISSING_FROM_SCHEMA = {'payments': {'capabilities', 'version_token', 'approved_money'}}
+        MISSING_FROM_SCHEMA = {'payments': {'capabilities', 'version_token', 'approved_money', 'refund_ids', 'refunded_money', 'processing_fee'}}
 
 
         # Loop first_sync_records and compare against second_sync_records